requirements.txt
Django>=2.1
xxhash
django-imagekit
django-appconf
Pillow
